Add MQTT Client to the MQTTBridge SIM: https://sim.amazon.com/issues/P39017723
This change adds an MQTTClient class
The MQTTClient uses the Topic mapping to subscribe to the required topics and also listens to the messages coming from the MessageBridge.
It also listens to changes in the Mapping and updates the subscriptions and internal mapping based on that.

The internal mapping can be further optimized for quick retrieval.

For the MQTTBridgeTest, I am spawning a moquette server.