Moving routing to MessageBridge
Decided to move the logic to MessageBridge as it was being duplicated in all the clients.
The clients are added to the MessageBridge. MessageBridge tracks the mapping and asks the clients to subscribe to topics and callback when a message is received. On getting a message callback, the bridge routes the message to other clients.