Add MQTT Client to the MQTTBridge (#6)
* Add MQTT Client to the MQTTBridge
SIM: https://sim.amazon.com/issues/P39017723

This change adds an MQTTClient class
The MQTTClient uses the Topic mapping to subscribe to the required topics and also listens to the messages coming from the MessageBridge.
It also listens to changes in the Mapping and updates the subscriptions and internal mapping based on that.

The internal mapping can be further optimized for quick retrieval.

For the MQTTBridgeTest, I am spawning a moquette server.

* Moving routing to MessageBridge

Decided to move the logic to MessageBridge as it was being duplicated in all the clients.
The clients are added to the MessageBridge. MessageBridge tracks the mapping and asks the clients to subscribe to topics and callback when a message is received. On getting a message callback, the bridge routes the message to other clients.

* Addressing PR comments.

* Addressing PR comments.

* Resubscribe on reconnect

* Address a PR comment