fix: do not recreate MQTTClient on service start
Previously, an MQTTClient was recreated every time the MQTTBridge
service started. Since the MQTTClientKeyStore held a reference to the
MQTTClient object, this resulted in multiple MQTTClient objects, each of
which attempted to connect when the key store updated. This led to
duplicate client ID issues in the MQTT broker.

This change ensures that only a single MQTTClient ever gets created.