feat: add support for topic filters in mapping. (#37)
This change adds support for topic filters for IotCore and LocalMqtt
sources. When a message arrives from one of these two sources, the
topic is matched against all possible topic filters and forwarded to the
matching destinations. Topics are forwarded to the topic that they are
received on.

PubSub does not support topic filters.