fix: add retry logic to initial mqtt connect (#39)
This change adds retry logic to the MQTT connect flows in the MQTTClient.
This connect flow executes whenever the component first starts, or any
time configuration (or key) is updated. It also re-writes MQTTClient unit tests
to utilize a FakeMqttClient rather than a mock. This greatly improves test
maintainability.

This change is required to prevent the component entering the BROKEN state
upon initial deployment if the MQTT broker is not yet available.