fix: remove client certificate rotate callback (#64)
It is not necessary to disconnect the MQTT Bridge when the client
certificate rotates. This change removes that logic, but still ensures
that the new client certificate is used the next time the bridge
connects.

CA rotation still results in a disconnect, as that indicates that the
client should no longer trust the server.